backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Andrey Chernikov"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.95.1"
uvicorn = "^0.21.1"
passlib = "^1.7.4"
tortoise-orm = "^0.16"
python-dotenv = "^1.0.0"
pydantic = "^1.10.7"
psycopg2-binary = "^2.9.6"
psycopg-binary = "^3.1.8"
psycopg-pool = "^3.1.7"
aerich = "^0.5"
asyncpg = "^0.27.0"

[tool.poetry.group.dev.dependencies]
isort = "^5.12.0"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
